refactor: RestorePage directory
