+++ src/view/components/RestorePage/StoredWindows.tsx
